Organize camera and viewports handlers/connectors
