+++ src/js/camera.test.js
@@ -6,54 +6,16 @@ import {
   getModelViewMatrix,
   getProjectionMatrix,
   getMVPMatrix,
-  getViewportAspect,
-  getViewportRect,
-  getViewportMatrix,
-  setViewportRect,
   moveBy,
   moveTo,
   zoomIn,
   zoomOut,
 } from "./camera";
-import { mat4, vec3 } from "gl-matrix";
-
-expect.extend({
-  toEqualMat4(received, expected) {
-    const pass = mat4.exactEquals(received, expected);
-    if (pass) {
-      return {
-        message: () =>
-          `expected [${received}] not to be equal to [${expected}]`,
-        pass: true,
-      };
-    } else {
-      return {
-        message: () => `expected [${received}] to be equal to [${expected}]`,
-        pass: false,
-      };
-    }
-  },
-  toEqualVec3(received, expected) {
-    const pass = vec3.exactEquals(received, expected);
-    if (pass) {
-      return {
-        message: () =>
-          `expected [${received}] not to be equal to [${expected}]`,
-        pass: true,
-      };
-    } else {
-      return {
-        message: () => `expected [${received}] to be equal to [${expected}]`,
-        pass: false,
-      };
-    }
-  },
-});
 
 describe("camera", () => {
   it("gets position", () => {
     let position = getPosition({ camera: { position: [10, 20, 0] } });
-    expect(vec3.exactEquals(position, [10, 20, 0])).toBe(true);
+    expect(position).toEqualVec3([10, 20, 0]);
   });
   it("gets scale", () => {
     let scale = getScale({ camera: { scale: 4 } });
@@ -179,32 +141,5 @@ describe("camera", () => {
     );
     expect(db.camera.position).toEqualVec3([75, 150, 0]);
     expect(db.camera.scale).toBe(100);
-  });
-});
-
-describe("viewport", () => {
-  it("gets rect", () => {
-    let rect = getViewportRect({ viewport: { rect: [0, 1, 2, 3] } });
-    expect(vec3.exactEquals(rect, [0, 1, 2, 3])).toBe(true);
-  });
-  it("gets aspect", () => {
-    let aspect = getViewportAspect([0, 1, 2, 3]);
-    expect(aspect).toBe(2 / 3);
-  });
-  it("gets viewport matrix", () => {
-    let viewport = getViewportMatrix([1, 2, 800, 600]);
-    expect(viewport).toEqualMat4(
-      // prettier-ignore
-      [
-        400,   0, 0, 0,
-          0, 300, 0, 0,
-          0,   0, 1, 0,
-        400, 300, 0, 1,
-      ],
-    );
-  });
-  it("sets rect", () => {
-    let db = setViewportRect({ viewport: { rect: [0, 0, 0, 0] } }, [10, 20, 30, 40]);
-    expect(db.viewport.rect).toEqual([10, 20, 30, 40]);
   });
 });
